md2conf-263 Add possibility to ignore local file updates when publishing
add skipUpdate to ConfluencePage

+++ confluence-client/src/main/java/io/github/md2conf/confluence/client/PublishConfluenceClient.java
@@ -201,6 +201,10 @@ public class PublishConfluenceClient {
         String newContentHash = hash(content);
 
         if (notSameHash(existingContentHash, newContentHash) || !existingPage.getTitle().equals(page.getTitle())) {
+            if (page.isSkipUpdate()){
+                this.publishConfluenceClientListener.pageSkippedUpdate(existingPage);
+                return;
+            }
             this.apiInternalClient.deletePropertyByKey(contentId, CONTENT_HASH_PROPERTY_KEY);
             int newPageVersion = existingPage.getVersion() + 1;
             this.apiInternalClient.updatePage(contentId, ancestorId, page.getTitle(), content, page.getType(), newPageVersion, this.versionMessage, this.notifyWatchers);
